Move discover_schema request to react-query
