+++ airbyte-webapp/src/core/domain/connector/types.ts
@@ -2,7 +2,10 @@ import {
   ConnectionConfiguration,
   ConnectionSpecification,
 } from "core/domain/connection";
-import { DestinationSyncMode } from "core/domain/catalog";
+import {
+  DestinationSyncMode,
+  SourceDiscoverSchemaRead,
+} from "core/domain/catalog";
 import { JobInfo } from "../job";
 
 export enum ReleaseStage {
@@ -127,4 +130,8 @@ export interface Scheduler {
   status: string;
   message: string;
   jobInfo?: JobInfo;
+}
+
+export interface Schema extends SourceDiscoverSchemaRead {
+  id: string;
 }
